wire: Move generic tests for perunio serializer
- Move from wire/test to wire/perunio/serializer/test.

- Because these generic tests are specific to perunio serializer.
