+++ wire/perunio/byteslice_external_test.go
@@ -22,14 +22,14 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"perun.network/go-perun/wire/perunio"
-	wiretest "perun.network/go-perun/wire/test"
+	peruniotest "perun.network/go-perun/wire/perunio/test"
 	ctxtest "polycry.pt/poly-go/context/test"
 )
 
 func TestByteSlice(t *testing.T) {
 	var v1, v2, v3, v4 perunio.ByteSlice = []byte{}, []byte{255}, []byte{1, 2, 3, 4, 5, 6}, make([]byte, 20000)
 	testByteSlices(t, v1, v2, v3, v4)
-	wiretest.GenericBrokenPipeTest(t, &v1, &v2, &v3, &v4)
+	peruniotest.GenericBrokenPipeTest(t, &v1, &v2, &v3, &v4)
 }
 
 // TestStutter tests what happens if the network stutters (split one message into several network packages).
